@@ -1,20 +1,18 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
+    <PropertyGroup>
+        <TargetFramework>net7.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+    </PropertyGroup>
 
-  <ItemGroup>
-    <PackageReference Include="Blackbird.Applications.Sdk.Common" Version="2.0.4" />
-    <PackageReference Include="Microsoft.AspNetCore.WebUtilities" Version="2.2.0" />
-    <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
-    <PackageReference Include="RestSharp" Version="110.1.0" />
-  </ItemGroup>
+    <ItemGroup>
+        <PackageReference Include="Blackbird.Applications.Sdk.Utils" Version="1.0.3" />
+        <PackageReference Include="Blackbird.Applications.Sdk.Common" Version="2.1.0" />
+    </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="Models\Workspaces\Requests\" />
-  </ItemGroup>
+    <ItemGroup>
+        <Folder Include="Models\Workspaces\Requests\" />
+    </ItemGroup>
 
 </Project>
