@@ -1,60 +1,70 @@
-﻿using Apps.Asana.Dtos;
-using Apps.Translate5;
+﻿using Apps.Asana.Actions.Base;
+using Apps.Asana.Api;
+using Apps.Asana.Constants;
+using Apps.Asana.Dtos;
+using Apps.Asana.Dtos.Base;
 using Blackbird.Applications.Sdk.Common;
-using Blackbird.Applications.Sdk.Common.Authentication;
 using RestSharp;
 using Apps.Asana.Models.Users.Requests;
 using Apps.Asana.Models.Users.Responses;
 using Blackbird.Applications.Sdk.Common.Actions;
+using Blackbird.Applications.Sdk.Common.Invocation;
+using Blackbird.Applications.Sdk.Utils.Extensions.String;
 
-namespace Apps.Asana.Actions
+namespace Apps.Asana.Actions;
+
+[ActionList]
+public class UserActions : AsanaActions
 {
-    [ActionList]
-    public class UserActions
+    public UserActions(InvocationContext invocationContext) : base(invocationContext)
     {
-        [Action("Get user", Description = "Get user by Id")]
-        public GetUserResponse GetUser(IEnumerable<AuthenticationCredentialsProvider> authenticationCredentialsProviders,
-           [ActionParameter] GetUserRequest input)
-        {
-            var client = new AsanaClient();
-            var request = new AsanaRequest($"/users/{input.UserId}", Method.Get, authenticationCredentialsProviders);
-            var user = client.Get<ResponseWrapper<UserDto>>(request);
-            return new GetUserResponse()
-            {
-                GId = user.Data.GId,
-                Name = user.Data.Name,
-                Email = user.Data.Email,
-                Workspaces = user.Data.Workspaces
-            };
-        }
+    }
+    
+    [Action("List users", Description = "List all users")]
+    public async Task<ListUsersResponse> ListUsers([ActionParameter] ListUsersRequest input)
+    {
+        var endpoint = ApiEndpoints.Users.WithQuery(input);
+        var request = new AsanaRequest(endpoint, Method.Get, Creds);
 
-        [Action("Get user's task list", Description = "Get user's task list by user Id")]
-        public GetUserTaskListResponse GetUserTaskList(IEnumerable<AuthenticationCredentialsProvider> authenticationCredentialsProviders,
-           [ActionParameter] GetUserTaskListRequest input)
-        {
-            var client = new AsanaClient();
-            var request = new AsanaRequest($"/users/{input.UserId}/user_task_list?workspace={input.WorkspaceId}", 
-                Method.Get, authenticationCredentialsProviders);
-            var userTaskList = client.Get<ResponseWrapper<UserTaskListDto>>(request);
-            return new GetUserTaskListResponse()
-            {
-                Id = userTaskList.Data.GId,
-                Name = userTaskList.Data.Name
-            };
-        }
+        var tasks = await Client.ExecuteWithErrorHandling<IEnumerable<AsanaEntity>>(request);
 
-        [Action("Get user's teams", Description = "Get user's teams")]
-        public GetUserTeamsResponse GetUserTeams(IEnumerable<AuthenticationCredentialsProvider> authenticationCredentialsProviders,
-           [ActionParameter] GetUserTeamsRequest input)
+        return new()
         {
-            var client = new AsanaClient();
-            var request = new AsanaRequest($"/users/{input.UserId}/teams?organization={input.WorkspaceId}",
-                Method.Get, authenticationCredentialsProviders);
-            var teams = client.Get<ResponseWrapper<List<WorkspaceDto>>>(request);
-            return new GetUserTeamsResponse()
-            {
-                Teams = teams.Data
-            };
-        }
+            Users = tasks
+        };
+    }
+
+    [Action("Get user", Description = "Get user by ID")]
+    public Task<UserDto> GetUser([ActionParameter] UserRequest input)
+    {
+        var endpoint = $"{ApiEndpoints.Users}/{input.UserId}";
+        var request = new AsanaRequest(endpoint, Method.Get, Creds);
+
+        return Client.ExecuteWithErrorHandling<UserDto>(request);
+    }
+
+    [Action("Get user's task list", Description = "Get user's task list by user ID")]
+    public Task<AsanaEntity> GetUserTaskList([ActionParameter] GetUserItemsRequest input)
+    {
+        var endpoint = $"{ApiEndpoints.Users}/{input.UserId}/user_task_list"
+            .SetQueryParameter("workspace", input.WorkspaceId);
+        var request = new AsanaRequest(endpoint, Method.Get, Creds);
+
+        return Client.ExecuteWithErrorHandling<AsanaEntity>(request);
+    }
+
+    [Action("Get user's teams", Description = "Get all user's teams")]
+    public async Task<GetUserTeamsResponse> GetUserTeams([ActionParameter] GetUserItemsRequest input)
+    {
+        var endpoint = $"{ApiEndpoints.Users}/{input.UserId}/teams"
+            .SetQueryParameter("workspace", input.WorkspaceId);
+        var request = new AsanaRequest(endpoint, Method.Get, Creds);
+        
+        var teams = await Client.ExecuteWithErrorHandling<IEnumerable<AsanaEntity>>(request);
+        
+        return new()
+        {
+            Teams = teams
+        };
     }
 }
