@@ -1,25 +1,42 @@
-﻿using Apps.Asana.Dtos;
-using Apps.Translate5;
-using Blackbird.Applications.Sdk.Common.Authentication;
+﻿using Apps.Asana.Actions.Base;
+using Apps.Asana.Api;
+using Apps.Asana.Constants;
+using Apps.Asana.Dtos;
+using Apps.Asana.Dtos.Base;
+using Apps.Asana.Models.Workspaces.Requests;
 using RestSharp;
 using Apps.Asana.Models.Workspaces.Responses;
+using Blackbird.Applications.Sdk.Common;
 using Blackbird.Applications.Sdk.Common.Actions;
+using Blackbird.Applications.Sdk.Common.Invocation;
 
-namespace Apps.Asana.Actions
+namespace Apps.Asana.Actions;
+
+[ActionList]
+public class WorkspaceActions : AsanaActions
 {
-    [ActionList]
-    public class WorkspaceActions
+    public WorkspaceActions(InvocationContext invocationContext) : base(invocationContext)
     {
-        [Action("List workspaces", Description = "List workspaces")]
-        public ListWorkspacesResponse ListAllWorkspaces(IEnumerable<AuthenticationCredentialsProvider> authenticationCredentialsProviders)
+    }
+
+    [Action("List workspaces", Description = "List workspaces")]
+    public async Task<ListWorkspacesResponse> ListAllWorkspaces()
+    {
+        var request = new AsanaRequest(ApiEndpoints.Workspaces, Method.Get, Creds);
+        var workspaces = await Client.ExecuteWithErrorHandling<IEnumerable<AsanaEntity>>(request);
+
+        return new()
         {
-            var client = new AsanaClient();
-            var request = new AsanaRequest($"/workspaces", Method.Get, authenticationCredentialsProviders);
-            var workspaces = client.Get<ResponseWrapper<List<WorkspaceDto>>>(request);
-            return new ListWorkspacesResponse()
-            {
-                Workspaces = workspaces.Data
-            };
-        }
+            Workspaces = workspaces
+        };
+    }
+
+    [Action("Get workspace", Description = "Get details of specific workspaces")]
+    public Task<WorkspaceDto> GetWorkspace([ActionParameter] WorkspaceRequest workspace)
+    {
+        var endpoint = $"{ApiEndpoints.Workspaces}/{workspace.WorkspaceId}";
+        var request = new AsanaRequest(endpoint, Method.Get, Creds);
+        
+        return Client.ExecuteWithErrorHandling<WorkspaceDto>(request);
     }
 }
