@@ -1,22 +1,22 @@
-﻿using Blackbird.Applications.Sdk.Common.Authentication.OAuth2;
-using Microsoft.AspNetCore.WebUtilities;
+﻿using Apps.Asana.Constants;
+using Blackbird.Applications.Sdk.Common.Authentication.OAuth2;
+using Blackbird.Applications.Sdk.Utils.Extensions.String;
 
-namespace Apps.Asana.Auth.OAuth2
+namespace Apps.Asana.Auth;
+
+public class OAuth2AuthorizeService : IOAuth2AuthorizeService
 {
-    public class OAuth2AuthorizeService : IOAuth2AuthorizeService
+    public string GetAuthorizationUrl(Dictionary<string, string> values)
     {
-        public string GetAuthorizationUrl(Dictionary<string, string> values)
+        var parameters = new Dictionary<string, string>
         {
-            const string oauthUrl = "https://app.asana.com/-/oauth_authorize";
-            var parameters = new Dictionary<string, string>
-            {
-                { "client_id", values["client_id"] },
-                { "redirect_uri", values["redirect_uri"] },
-                { "response_type", values["response_type"] },
-                { "scope", values["scope"] },
-                { "state", values["state"] }
-            };
-            return QueryHelpers.AddQueryString(oauthUrl, parameters);
-        }
+            { "client_id", values[CredsNames.ClientId] },
+            { "redirect_uri", values[CredsNames.RedirectUri] },
+            { "response_type", values[CredsNames.ResponseType] },
+            { "scope", values[CredsNames.Scope] },
+            { "state", values["state"] }
+        };
+        
+        return Urls.OAuthUrl.WithQuery(parameters);
     }
 }
