@@ -1,70 +1,79 @@
-﻿using Blackbird.Applications.Sdk.Common.Authentication.OAuth2;
-using System.Text.Json;
+﻿using System.Globalization;
+using Apps.Asana.Constants;
+using Blackbird.Applications.Sdk.Common.Authentication.OAuth2;
+using Newtonsoft.Json;
 
-namespace Apps.Asana.Auth.OAuth2
+namespace Apps.Asana.Auth;
+
+public class OAuth2TokenService : IOAuth2TokenService
 {
-    public class OAuth2TokenService : IOAuth2TokenService
+    private const string ExpiresAtKeyName = "expires_at";
+
+    public bool IsRefreshToken(Dictionary<string, string> values)
     {
-        private const string ExpiresAtKeyName = "expires_at";
-        private const string TokenUrl = "https://app.asana.com/-/oauth_token";
+        var expiresAt = DateTime.Parse(values[ExpiresAtKeyName]);
+        return DateTime.UtcNow > expiresAt;
+    }
 
-        public bool IsRefreshToken(Dictionary<string, string> values)
+    public Task<Dictionary<string, string>> RefreshToken(Dictionary<string, string> values,
+        CancellationToken cancellationToken)
+    {
+        var bodyParameters = new Dictionary<string, string>
         {
-            var expiresAt = DateTime.Parse(values[ExpiresAtKeyName]);
-            return DateTime.UtcNow > expiresAt;
-        }
+            { "grant_type", "refresh_token" },
+            { "client_id", values[CredsNames.ClientId] },
+            { "client_secret", values[CredsNames.ClientSecret] },
+            { "refresh_token", values["refresh_token"] },
+        };
+        return GetToken(bodyParameters, cancellationToken);
+    }
 
-        public async Task<Dictionary<string, string>> RefreshToken(Dictionary<string, string> values, CancellationToken cancellationToken)
+    public Task<Dictionary<string, string>> RequestToken(
+        string state,
+        string code,
+        Dictionary<string, string> values,
+        CancellationToken cancellationToken)
+    {
+        var bodyParameters = new Dictionary<string, string>
         {
-            const string grant_type = "refresh_token";
+            { "grant_type", "authorization_code" },
+            { "client_id", values[CredsNames.ClientId] },
+            { "client_secret", values[CredsNames.ClientSecret] },
+            { "redirect_uri", values[CredsNames.RedirectUri] },
+            { "code", code }
+        };
+        return GetToken(bodyParameters, cancellationToken);
+    }
 
-            var bodyParameters = new Dictionary<string, string>
-            {
-                { "grant_type", grant_type },
-                { "client_id", values["client_id"] },
-                { "client_secret", values["client_secret"] },
-                { "refresh_token", values["refresh_token"] },
-            };
-            return await RequestToken(bodyParameters, cancellationToken);
-        }
+    public Task RevokeToken(Dictionary<string, string> values)
+    {
+        throw new NotImplementedException();
+    }
 
-        public async Task<Dictionary<string, string?>> RequestToken(
-            string state,
-            string code,
-            Dictionary<string, string> values,
-            CancellationToken cancellationToken)
-        {
-            const string grant_type = "authorization_code";
+    private async Task<Dictionary<string, string>> GetToken(Dictionary<string, string> parameters,
+        CancellationToken token)
+    {
+        var responseContent = await ExecuteTokenRequest(parameters, token);
 
-            var bodyParameters = new Dictionary<string, string>
-            {
-                { "grant_type", grant_type },
-                { "client_id", values["client_id"] },
-                { "client_secret", values["client_secret"] },
-                { "redirect_uri", values["redirect_uri"] },
-                { "code", code }
-            };
-            return await RequestToken(bodyParameters, cancellationToken);
-        }
+        var resultDictionary = JsonConvert.DeserializeObject<Dictionary<string, string>>(responseContent)
+                                   ?.ToDictionary(r => r.Key, r => r.Value.ToString())
+                               ?? throw new InvalidOperationException(
+                                   $"Invalid response content: {responseContent}");
 
-        public Task RevokeToken(Dictionary<string, string> values)
-        {
-            throw new NotImplementedException();
-        }
+        var expiresIn = int.Parse(resultDictionary["expires_in"]);
+        var expiresAt = DateTime.UtcNow.AddSeconds(expiresIn);
+        resultDictionary.Add(ExpiresAtKeyName, expiresAt.ToString(CultureInfo.InvariantCulture));
 
-        private async Task<Dictionary<string, string>> RequestToken(Dictionary<string, string> bodyParameters, CancellationToken cancellationToken)
-        {
-            var utcNow = DateTime.UtcNow;
-            using HttpClient httpClient = new HttpClient();
-            using var httpContent = new FormUrlEncodedContent(bodyParameters);
-            using var response = await httpClient.PostAsync(TokenUrl, httpContent, cancellationToken);
-            var responseContent = await response.Content.ReadAsStringAsync();
-            var resultDictionary = JsonSerializer.Deserialize<Dictionary<string, object>>(responseContent)?.ToDictionary(r => r.Key, r => r.Value?.ToString())
-                ?? throw new InvalidOperationException($"Invalid response content: {responseContent}");
-            var expriresIn = int.Parse(resultDictionary["expires_in"]);
-            var expiresAt = utcNow.AddSeconds(expriresIn);
-            resultDictionary.Add(ExpiresAtKeyName, expiresAt.ToString());
-            return resultDictionary;
-        }
+        return resultDictionary;
+    }
+
+    private async Task<string> ExecuteTokenRequest(Dictionary<string, string> parameters,
+        CancellationToken cancellationToken)
+    {
+        using var client = new HttpClient();
+        using var content = new FormUrlEncodedContent(parameters);
+        using var response = await client.PostAsync(Urls.TokenUrl, content, cancellationToken);
+
+        return await response.Content.ReadAsStringAsync(cancellationToken);
     }
 }
