@@ -1,54 +1,54 @@
-﻿using Blackbird.Applications.Sdk.Common.Authentication;
+﻿using Apps.Asana.Constants;
+using Blackbird.Applications.Sdk.Common.Authentication;
 using Blackbird.Applications.Sdk.Common.Connections;
 
-namespace Apps.Asana.Connections
+namespace Apps.Asana.Connections;
+
+public class ConnectionDefinition : IConnectionDefinition
 {
-    public class ConnectionDefinition : IConnectionDefinition
+    public IEnumerable<ConnectionPropertyGroup> ConnectionPropertyGroups => new List<ConnectionPropertyGroup>
     {
-
-        public IEnumerable<ConnectionPropertyGroup> ConnectionPropertyGroups => new List<ConnectionPropertyGroup>()
+        // new()
+        // {
+        //     Name = "OAuth2",
+        //     AuthenticationType = ConnectionAuthenticationType.OAuth2,
+        //     ConnectionUsage = ConnectionUsage.Actions,
+        //     ConnectionProperties = new List<ConnectionProperty>
+        //     {
+        //         new(CredsNames.ClientId) { DisplayName = "Client ID" },
+        //         new(CredsNames.ClientSecret) { DisplayName = "Client secret" },
+        //         new(CredsNames.RedirectUri) { DisplayName = "Redirect URI" },
+        //         new(CredsNames.Scope) { DisplayName = "Scope" },
+        //         new(CredsNames.ResponseType) { DisplayName = "Response type" },
+        //     }
+        // },
+        new()
         {
-            new ConnectionPropertyGroup
+            Name = "Developer API token",
+            AuthenticationType = ConnectionAuthenticationType.Undefined,
+            ConnectionUsage = ConnectionUsage.Webhooks,
+            ConnectionProperties = new List<ConnectionProperty>
             {
-                Name = "OAuth2",
-                AuthenticationType = ConnectionAuthenticationType.OAuth2,
-                ConnectionUsage = ConnectionUsage.Actions,
-                ConnectionProperties = new List<ConnectionProperty>()
-                {
-                    new ConnectionProperty("client_id"),
-                    new ConnectionProperty("client_secret"),
-                    new ConnectionProperty("redirect_uri"),
-                    new ConnectionProperty("scope"),
-                    new ConnectionProperty("response_type"),
-                }
-            },
-            new ConnectionPropertyGroup
-            {
-                Name = "Developer API token",
-                AuthenticationType = ConnectionAuthenticationType.Undefined,
-                ConnectionUsage = ConnectionUsage.Webhooks,
-                ConnectionProperties = new List<ConnectionProperty>()
-                {
-                    new ConnectionProperty("apiToken")
-                }
+                new(CredsNames.ApiToken) { DisplayName = "API token" }
             }
-        };
-
-        public IEnumerable<AuthenticationCredentialsProvider> CreateAuthorizationCredentialsProviders(Dictionary<string, string> values)
-        {
-            var token = values.First(v => v.Key == "access_token");
-            yield return new AuthenticationCredentialsProvider(
-                AuthenticationCredentialsRequestLocation.Header,
-                "Authorization",
-                $"Bearer {token.Value}"
-            );
-
-            var apiToken = values.First(v => v.Key == "apiToken");
-            yield return new AuthenticationCredentialsProvider(
-                AuthenticationCredentialsRequestLocation.None,
-                apiToken.Key,
-                apiToken.Value
-            );
         }
+    };
+
+    public IEnumerable<AuthenticationCredentialsProvider> CreateAuthorizationCredentialsProviders(
+        Dictionary<string, string> values)
+    {
+        // var accessToken = values.First(v => v.Key == CredsNames.AccessToken);
+        // yield return new AuthenticationCredentialsProvider(
+        //     AuthenticationCredentialsRequestLocation.Header,
+        //     accessToken.Key,
+        //     accessToken.Value
+        // );
+
+        var apiToken = values.First(v => v.Key == CredsNames.ApiToken);
+        yield return new AuthenticationCredentialsProvider(
+            AuthenticationCredentialsRequestLocation.None,
+            apiToken.Key,
+            apiToken.Value
+        );
     }
 }
