@@ -1,14 +1,8 @@
-﻿using Apps.Asana.Dtos;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using Apps.Asana.Dtos.Base;
 
-namespace Apps.Asana.Models.Attachments.Responses
+namespace Apps.Asana.Models.Attachments.Responses;
+
+public class GetAttachmentsResponse
 {
-    public class GetAttachmentsResponse
-    {
-       public IEnumerable<AttachmentDto> Attachments { get; set; }
-    }
+    public IEnumerable<AsanaEntity> Attachments { get; set; }
 }
