@@ -1,15 +1,25 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using Apps.Asana.DataSourceHandlers;
+using Apps.Asana.DataSourceHandlers.EnumDataHandlers;
+using Blackbird.Applications.Sdk.Common;
+using Blackbird.Applications.Sdk.Common.Dynamic;
 
-namespace Apps.Asana.Models.Projects.Requests
+namespace Apps.Asana.Models.Projects.Requests;
+
+public class CreateProjectRequest
 {
-    public class CreateProjectRequest
-    {
-        public string TeamId { get; set; }
+    [Display("Team ID")] public string Team { get; set; }
 
-        public string ProjectName { get; set; }
-    }
+    [Display("Project name")] public string Name { get; set; }
+
+    [Display("Workspace")] 
+    [DataSource(typeof(WorkspaceDataHandler))]
+    public string? Workspace { get; set; }
+
+    [Display("Is public")] public bool? Public { get; set; }
+
+    public string? Owner { get; set; }
+    
+    [DataSource(typeof(ColorDataHandler))]
+    public string? Color { get; set; }
+    public bool? Archived { get; set; }
 }
