@@ -1,12 +1,21 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using Apps.Asana.DataSourceHandlers;
+using Blackbird.Applications.Sdk.Common;
+using Blackbird.Applications.Sdk.Common.Dynamic;
+using Newtonsoft.Json;
 
-namespace Apps.Asana.Models.Projects.Responses
+namespace Apps.Asana.Models.Projects.Requests;
+
+public class ListProjectsRequest
 {
-    public class ListProjectsRequest
-    {
-    }
+    [JsonProperty("workspace")]
+    [Display("Workspace")]
+    [DataSource(typeof(WorkspaceDataHandler))]
+    public string? Workspace { get; set; }
+    
+    [JsonProperty("team")]
+    [Display("Team ID")]
+    public string? Team { get; set; }
+    
+    [JsonProperty("archived")]
+    public bool? Archived { get; set; }
 }
