@@ -1,17 +1,13 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using Apps.Asana.DataSourceHandlers.EnumDataHandlers;
+using Blackbird.Applications.Sdk.Common.Dynamic;
 
-namespace Apps.Asana.Models.Tags.Requests
+namespace Apps.Asana.Models.Tags.Requests;
+
+public class UpdateTagRequest
 {
-    public class UpdateTagRequest
-    {
-        public string TagId { get; set; }
-
-        public string NewName { get; set; }
-
-        public string NewColor { get; set; }
-    }
+    public string? Name { get; set; }
+    
+    [DataSource(typeof(ColorDataHandler))]
+    public string? Color { get; set; }
+    public string? Notes { get; set; }
 }
