@@ -1,14 +1,8 @@
-﻿using Apps.Asana.Dtos;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using Apps.Asana.Dtos.Base;
 
-namespace Apps.Asana.Models.Tags.Responses
+namespace Apps.Asana.Models.Tags.Responses;
+
+public class ListTagsResponse
 {
-    public class ListTagsResponse
-    {
-        public IEnumerable<TagDto> Tags { get; set; }
-    }
+    public IEnumerable<AsanaEntity> Tags { get; set; }
 }
