@@ -1,15 +1,12 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using Apps.Asana.DataSourceHandlers;
+using Blackbird.Applications.Sdk.Common;
+using Blackbird.Applications.Sdk.Common.Dynamic;
 
-namespace Apps.Asana.Models.Tasks.Requests
+namespace Apps.Asana.Models.Tasks.Requests;
+
+public class AssignTagRequest
 {
-    public class AssignTagRequest
-    {
-        public string TagId { get; set; }
-
-        public string TaskId { get; set; }
-    }
+    [Display("Tag")]
+    [DataSource(typeof(TagDataHandler))]
+    public string Tag { get; set; }
 }
