@@ -1,13 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿namespace Apps.Asana.Dtos;
 
-namespace Apps.Asana.Dtos
+public class ProjectStatusUpdateDto
 {
-    public class ProjectStatusUpdateDto
-    {
-        public string Title { get; set; }
-    }
+    public string Title { get; set; }
 }
